FailMail, Large donation test, other tests testing mails - fix to use MailFactory
This removes one live place where the deprecated Mailer class
is used & all the test instances.

Per https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/878237/1/drupal/sites/all/modules/wmf_campaigns/wmf_campaigns.module#272

We find that we need to change 'to' to to_address for the
SMTP mailer class (currently used for thank you mail etc)

This also fixes all the test places

Note - I find the large donation test is not running well locally
in isolation before or after
as the hooks are not loading consistently from the
drupal class. My take is that they will be
moved out of the drupal class soon so I guess if this is
no worse on jenkins it makes sense to merge

Bug: T326271

Change-Id: Ibf0562b0063bd25ea7dc2ba19178d229a8b9b2d4