Make pending txn resolver work for PayPal
This patch just makes it work for one-time donations - we can add
recurring in a follow-on.

The gateway_txn_id is now added directly to the message rather than
passed along in a separate variable. We also fix some test setup
so that it reflects reality - Adyen's getLatestPaymentStatus call
will always return the same gateway_txn_id as was in the original
message, and the approvePayment call will return a different ID.

We also break out one more function to send the donation message.

The PayPal resolve tests are pretty basic, just making sure the
new info from the status lookup gets into the right place. Since
the existing donor lookup is the same across all the gateways I
didn't feel like we needed to test it again here.

The best way I found to make a resolvable orphan was to stop the
docker containers after the redirect to paypal, click the 'Complete
Purchase' button at paypal and get the error at my port 9001, then
start them back up again and consume the queues.

You can run the job with a zero minimum age so you don't need to
wait to resolve the pending message:

  drush @wmff -vv cvapi PendingTable.consume version=4 \
    gateway=paypal_ec timeLimit=1500 minimumAge=0

Bug: T324642
Change-Id: I59bbefb16562116ae1bcb1d3a9f46ed17f8bfd18