Some test cleanups to support updating phpunit
I tried to update phpunit but we have some uncompatible
code in our tests. This fixes some.

Note in phpunit.xml we are referencing test that have been
moved to extensions & that upset the later phpunit.

Also, my IDE replaced the list functions on save - but
I think we might as well change that

Change-Id: I36efcfaa88918c5c3bc6c4ebbde5343eab45a0d9