Add country to recurring charge calls
It's required for dlocal recurring charges. Note that the contact
address needs to match the country on the charge in the frontend.
If you're testing with e.g. Jimmy your frontend charge might have
been merged into an existing contact with a different country, so
you'll have to update the country manually in the Civi UI.

We also switch the contact lookup to use API4 to get the ISO code
more easily.

Bug: T324298
Change-Id: I633901734d91258b2a8464629a79cdfc58df3c2c