Duck failure
This is deeply confusing and took me a long time to replicate...
The test setup does replicate

ie create 2 contacts. Give each a contribution, the contact with the
lower contact id should have the more recent contribution.

Give each an address. It is important that the address has a conflicting postal code.
Merge them through the deduper UI - the correct behaviour is that the address
from the contact who has the lower contact contact id and the more
recent contribution is kept.

This bug arises from some very nasty code which is complicated because it has to mess
with a very confusing array of values. However, what seemed to be going on is
that is is adding in is_primary as a field to trigger an over-write, regardless of
whether is is the same between the 2 addresses

We do have a lot of tests on this handling, mostly cos it is so dicey, here & in the deduper extension

Bug: T330231

Change-Id: Id4c93185296ed41bf9642f9bdf646388082e581b