Exclude contribution tracking from log tables
This is a replacement for
https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/908584

The table wasn't being excluded in ^^ because the purpose of the line
adding wmf_contribution_extra was to ADD it to the logged tables
rather than remove it. Clearly code comments were missing!

However, I tested & that line is no longer required to add it in - because
the list of tables actually checks civicrm_custom_group.table_name
now rather than using a string match on the civicrm_ prefix - so I removed
it to make the code clearer

Bug: T334533
Change-Id: I65a76e6060fbd5d98218d7a3480ca221c3fbc647