Re-add payments_form field, as 3 separate fields
It looks like we omitted payments_form in the new entity and
that appears to not be deliberate.

This re-adds it, broken down into 3 component fields
- gateway
- variant
- appeal

As we have deployed the field but not started using it I incremented the version
on the upgrader and did a DROP TABLE + the original create.

Bug: T334532
Change-Id: I818e4d90e383c17b8a899b353f1f1eeb14b4cb61