Additional contribution_tracking fields proposed by analytics
In discussion with Joseph this is the additional data analytics pulls out
of utm_key into meaningful fields. In addition he pointed out the risk
not retaining utm_source in addition to the break out fields - ie that
it was subject to user error and we would lose the ability to
see what they entered if they made mistakes.

I have added this per the fields in
https://docs.google.com/document/d/1oxtIfTHNwOzmsiogVZ9KkA-ghVLNsKKMiReSBJoGadQ/edit

However, after discussing with Elliot I have
used boolean or integer fields (with option lists) where possible

Bug: T334532
Change-Id: I4d8b5a2c5a08c8638f39b03cb67a4050c5edc523