Populate contribution tracking fields
To run  drush @wmff civicrm-populate-contribution-tracking

-optionally add -vv to see the query
and --batch=15000 to alter the back size

This is the population script converted to drush. Note that as a throw away script
I'm happy for it to live in drupal & now have a clear separation of concerns re
the sql being in the same file. That also makes the db name wrangling easier.

Note the payment submethods won't go in right unless we fix the name on them in the db.
e.g on staging I got one working with

$results = \Civi\Api4\OptionValue::update()
  ->addValue('name', 'upi')
  ->addWhere('option_group_id:name', '=', 'payment_instrument')
  ->addWhere('value', '=', 238)
  ->execute();
- also 235 to netbanking and 236 to paytmwallet
- and 9 to rtbt_ideal

Bug: T334534
Change-Id: I44fd0d0d623ea7267f6e67e2066f0aa76a7633be