Fill scheme ID for ingenico recurrings
Uses an API call that we have wired up in SmashPig without any of
the normalization. That's OK, we can just check the two possible
spots in the raw response.

You can try running the same API call in production with a SmashPig
maintenance script:

php GetPaymentStatus.php --id=000000657070022735910000200001

We can run this in batches till we see it stops returning anything
in the results. Then we can go back and delete all the custom value
rows we have inserted with blank strings in the scheme ID.

Bug: T331754
Change-Id: Ib25e3f58780d3a6944cd763ef7f846bc45008eee