Fix handling where duplicate contacts being chosen between
This addresses an issue Melanie had where there were 4 duplicate contacts relating to a
matching gift. Two had employer relationships & 2 had prior soft credit connections.
The right fix was to merge them - meaning the right approach in the import
was to give a meaningful error (not a typeHint...).

I added an escape valve for the situation where there really are 2
contacts called Jane Doe, both of which have a matching gift
history with The Firm but only one of which is still employed - in
this case ensuring one has an inactive relationship will mark it
as non-duplicate. I don't know this could ever arise but adding
handling means there is 'something' we could tell the user to do.

Note this is stricter than Benevity - I think this reflects our gradually
improving data but I didn't want to impose it on Benevity
as the fix & re-run is way harder there & the volume is higher so
I left that as 'non-strict'

Bug: T336609
Change-Id: Icb6c63558b34020f12077db756b6fa26cd553875