Move our transaction handling class to extension
I hit the old drupla namespace issues on the WMFDatabase class in trying to
add a test. Rather than do require_once etc I decided to
bite the bullet & move it to our namespace structure.
The challenge is that we have calls to drupal
functions that we didn't want to lose
so I figured we could move them into
a new Drupal class so we could still keep them in one place

Change-Id: Ided3bdf956fc9ff1e3bfdc8889a6b17b45dcb009