Restructure CalculateData class such that the sql for fields is in the array declaration.
This is a fairly significan restructure with the intent to be able to access
parts of this calculation more selectively rather than a big
sql generation code chunk allows. This is a first step rather than the
full planned change but I was trying to achieve a reviewable chunk of change

It is pretty tough to read the code change but I was able to
generate what the code outputs so comparing that is an appropriate way to review this.

The end result of this code is
to generate triggers for the contribution table that populate wmf_donor.

I put up a commit demonstrating the 'before' output
https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/925143/8
and a second one showing how the 'after' output changes
https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/927320/1

In the course of doing this I found that there is
an existing bug + some minor changes which I have annotated

https: //gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/927320/1/drupal/sites/all/modules/wmf_civicrm/scripts/contribution_triggers_local.sql

Bug: T336305

Change-Id: Ic8763140677563c2a5f6f48a65988d98143edd6b