Add WMF Donor api & test
The goal here is both to help write the upgrade script & for us to check what is being
calculated during development.

The key functionality is that if debug=TRUE it outputs the sql - which has
been a bit tricky to obtain when running on staging.

This works with the command

echo '{"select":["last_donation_currency"],"where":[["id","=",5]],"version":4, "debug":1}' | drush @wmff cvapi WMFDonor.get --in=json

See https://wmff.localhost:32353/civicrm/api4#/explorer/WMFDonor/get?where=%5B%5B%22id%22,%22%3D%22,%225%22%5D%5D&select=%5B%22last_donation_currency%22%5D&limit=0
(although it does blow out the screen width - sorry not sorry)

This is an initial version which I expect to iterate  - notably I'm tinkering
with how best to run it to see what it generates and the results. However, if this
passes tests I think we should merge as-is as I got through a few challenges to
get this far & will iterate for for further fixes

Bug: T336305

Change-Id: Ie9b04546d3e601475d92a6ee38b7efc89430d860