Add segment calculation information to contact's contribution tab to support our QA efforts
This is intended as temporary for us & stakeholders to QA the calculation - it
adds a note to the contact contribution tab showing what WOULD be calculated.

It also provides a page showing the calculated segments.

This version is simpler than their criteria - in part because I'm having trouble understanding
some of the criteria -but mostly because I need to get enough functionality to switch
focus onto how we would actually run this - ie I need to test both one-off updates and
updates within triggers.

I'm hoping we can go live with this simpler version and iterate - but first step
towards that is giving this some visibility.

Bug: T336305

Change-Id: I5afe793370297b01ed4b67967e674b98750dfa67