Add field value to WMF donor list of statuses and segments
Also did a small cosmetic change to seperate the tables for readability.

To test:
Pull down this patch and visit http://wmff.localhost:32353/civicrm/wmf-segment

Each table should have a heading and a third column. The third column should be equal to the values in the getDonorStatusOptions
 and getDonorSegmentOptions in the CalculatedData.php class.

If this page fails to load, it means the wmf-civicrm may be disabled. To enable it, use:
cv enable wmf-civicrm

Bug: T339297
Change-Id: Ic9629d1b872228358de257553ba96f37c5fed350