Add label for non-donor status
This is what is causing those errors - because we are calculating 'demo' statuses
(for QA purposes) on the fly and are not handling timezones correctly it is
determining they are 'non donor's and we don't have a label for that.

Note that this whole screen is temporary / in flux as
the requirements settle so it seems fine just to fix the label
and not worry about the timezone calculation issue.

Also note I added the label on prod - but that is not what is being
used in this context

Note you can replicate this locally by accessing the contributions
tab of a contact whose only contribution is in the future

Bug: T339971

Change-Id: Ie8c12f8d4a0ed5c4f16383362be00118bbbe16cd