Follow up fix on fredge report
This is additional fixes for

https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/931265

Note that to smoke test I needed to give the read only user permissions
GRANT ALL PRIVILEGES ON fredge.* TO 'civiro'@'%' ;
GRANT ALL PRIVILEGES ON drupal.* TO 'civiro'@'%' ;

- ideally we would do this for the fredge db in our setup somewhere. Perhaps not needed
for drupal once this is merged.

I tested these urls
https://wmff.localhost:32353/civicrm/report/wmffraud/fraud
https://wmff.localhost:32353/civicrm/report/wmffraud/fredge
https://wmff.localhost:32353/civicrm/report/wmffraud/paymentattempts

and made sure the output was the same with & without. Note that there
are a tonne of notices - these can be ignored & are unaffected.
Also note I made one minor tidy up to remove the reference to Drupal
Bug: T339381

Change-Id: I34f488e732aeac2c014a8353feeb38d70c1b1b31