Fix early return on no-records found
There was code to return when no records were found, but it needed to be
updated to reflect the code switch to apiv4.

This would be hit when a contact has a contribution with no contribution_tracking_id entry.

Note I tested this on staging with the drush command for the contact in question -

 drush @wmff cvapi contact.showme id=31415709 debug=1

which failed without the patch & passes with it deployed

Bug: T341633
Change-Id: Ie8bf9b737c284bcafc7aff3f29f2a485edfb334b