Convert PaymentsFraud access to use apiv4
This is a follow on to the fix in this code. I found the drupal query swallowed the error
so I added support for PaymentsFraud as an api entity and used that

This can be tested by added a contribution with associated record
in civicrm_contribution_tracking AND fredge.payments_fraud to the local database.

Then the Contact.showme api can be called or it can be triggered via this
page
https://wmff.localhost:32353/civicrm/a/#/forgetme/forget/16

Where 16 is the contact id.

Search the results for the word 'Fredge'

or

drush @wmff cvapi Contact.showme id=16 | grep Fredge

I didn't fix all the places - but I tried to comment them.

Bug: T341633
Change-Id: I429362bb208b1e42f711214ff1e6516e6cdf5880