Improve logging for thank you mailing creating
On looking at https://phabricator.wikimedia.org/T340064
I realised the logged error is misleading - it gives that message before attempting
to create a mailing record, which it presumably does successfully.

This logs that it intends to and that it has created the record. Or course it might
turn out when we deploy it that the logging doesn't tell us that, and then
we have a pointer towards a further bug.

I also switched to logging the revision & name rather than the subject as
that is what is being used.

Bug: T340064
Change-Id: I556bb3dc7501e2f08d3858d41ad0f3cc7dbfc16e