Move CiviMailStore class from the module to the extension
This follows our pattern of slowly fixing these namespace issues as we touch code.

This code is pretty nasty & confusing - but easier to deal with when it
is not also in the psuedo-module namespace and the exceptions are
not jammed into the same file.

Note that we have specific test cover in the ThankYouTest file and specifically in
testSendThankYouAddCiviMailActivity

Apart from fixing name-spacing & formatting there are no other changes

Bug: T340064

Change-Id: Ie112f2ba9e2b6f6a8f846eba8c2c7201736f60f2