Follow up tweaks on wmf_audit patch
This builds on https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/919940/8
with the primary goal being to change the check on $result as it would
be checking $result['id'] on NULL where the row is not found.

But, When I changed that it was really obvious that the if should be combined
with the if($result) below so I did that, with a minor re-ordering to early-return
cos that seemed clearer. Of course it makes the patch hard to read with the white
space changes...

Lastly I added the lines to the test I used to check the no result scenario
and made peace with my spell-checker

Bug: T335711

Change-Id: If56970f85b332185658082aeb09dc0666873eb7a