Add row number to salt when generating a transaction ID for anonymous
We 'make up' a transaction ID when importing contributions from spreadsheets if
one is not provided. Where we have a check number this made up ID is the check
number and the name (with some hashing to get rid of special characters
and make the trxn_id not look like someones name). However,
these spreadsheets often have several rows for one organization
check against different
individuals and within that they sometimes have multiple rows
agains the 'anonymous donor'. We need to increase the
detail in the transactionID to permit the same check number
multiple times against the anonymous donor or an organization

Further fixes are really just to be able to simulate this anonymous user in tests
and to reflect the fact the first_name & last_name might be the word anonymous rather than
blank. I made some fixes towards improving cleanup (using repetitive names) - but it
is still leaving cruft

This code is called from both the old and new imports

To test the old import (which is what the ticket was logged against) use
the url

https://wmff.localhost:32353/admin/import_checks

You can edit the file drupal\sites\all\module\offline2civicrm\tests\data\engage_individual.csv
such that both contacts have first name & last name = Anonymous
& confirm that  both rows will import for 'Engage Import' with this edited file.
(you will need to delete contributions between attempts)
You also need to set Money Order Number to empty in that file as
it is mapped to transaction ID & will be used if present

Bug: T337894
Bug: T339050
Change-Id: Ifed5dedca5aaccf33f4909e51c49d6720fbe220a