CiviCRM 5.64 beta, latest, stock
This pulls in things that have been merged to the
upstream rc since we deployed and also reverts
the temporary patches we applied in order to get the
upgrade script to run better.

It is general housekeeping to keep grabbing the latest
patches for our point version. But we also
specifically want this patch
https://github.com/civicrm/civicrm-core/pull/26811 to
revert the loss of the start date fields

Patches merged to the rc since it was cut are here
https://github.com/civicrm/civicrm-core/pulls?q=is%3Apr+is%3Aclosed+label%3A5.64
(although some of those were already merged when we pushed it out).

Note this isn't in the sprint but it's better to get back to a
'stock' deployment before porting in the patch
that I'm actually work on - ie
https://github.com/civicrm/civicrm-core/pull/26867
Bug: T341877

Change-Id: Ib6532188782545950441a42413d8751cd04b503d