Update Segment and status calculation to most recent interpretation.
Note that the calculated values can be seen on
the contribution tab and in the triggers (if they are rebuilt).

The triggers populate wmf_donor. However, you may need to use
mysql to see that as there is a bug (unchanged by this patch)
around the creation of the option values for these fields.

Changes in this PR relate to

- updating  the calculation of donor segments to be based on the
last 5 financial years, plus the current one (so that all statuses transition on
July 1)
- changing the triggers to have the current year hard coded at the time
of creation. That means that as we roll over from
one financial year to the next the field will stop updating
until we reload the triggers.

This is as outlined in https://phabricator.wikimedia.org/T342165
and the linked google doc

and for statuses
I put up quite a long comment on the phab
https://phabricator.wikimedia.org/T342165#9030172

My goal is to keep the code up to date with
how we understand the spec.

Note that to see the status calculation in action
try creating a
recurring contribution against a contact (or
find on on staging) with a recurring payment within 36 months
of now. The status in the texty-bit on the contribution
tab should change if you alter the contact's contributions
so that they do or don't quality as major or mid tier donor

We can safely deploy this as it won't affect the trigger
updates currently happening unless we actively reload them
(and even then - we have to ditch & reload the data
that HAS been trigger populated again anyway).

Bug: T342165

Change-Id: I53f90f4d13d19d1a5d6126dd32a1db9460035ec2