Fix option label create bug
This addresses the e-notice we have been seeing when running civibuild.create

[PHP Notice] Undefined index: id at /src/wikimedia/fundraising/crm/drupal/sites/all/modules/civicrm/Civi/Api4/Action/CustomField/CustomFieldSaveTrait.php:53

and instead correctly creates the labels for our custom groups.

There are 2 ways to declare the option values for a custom field
- as a value=>label array, as a value => array of values array
We mostly do the first but donor segment & status are declared
as arrays (& probably we will move that way with the others). However
the code upstream now (since 2020) expects 'id' to be populated
with 'value' - so this does that. It also removes handling for the
first sort of array as core does that fine now.

The code has evolved a bit over time in core - in particular with
https://github.com/civicrm/civicrm-core/pull/17167

Note once a civibuild has been done both the new style & the old style have values
- ie donor segment & gift source

https://wmff.localhost:32353/civicrm/admin/custom/group/field/option?reset=1&action=browse&gid=8&fid=22
https://wmff.localhost:32353/civicrm/admin/custom/group/field/option?reset=1&action=browse&gid=11&fid=164

Bug: T342943
Change-Id: Id637fd75a42faff4b6ee525422ef2965984b37cb