Add code to set snooze
To test this it is necessary to configure the sandbox database in your local
dev & probably do any testing over the VPN. Refer to
https://wikitech.wikimedia.org/wiki/Fundraising/Data_and_Integrated_Processes/Acoustic_Integration#Sandbox

It's hard to test this if you can't log into the acoustic UI but
at a pinch you can use the prod user name & password in the UI login
at https://login.goacoustic.com/signin

The code is configured to kick in when the new email custom field snooze_date
is edited. Note I have been testing that by creating an editable search on
email addresses in SearchKit. I have packaged is as a follow on commit
- if you load that commit & do a managed reconcile (or rebuild)
you will get the display at https://wmff.localhost:32353/civicrm/search#/display/snooze/snooze_table

If you edit a record and add a snooze date in the future it will queue an update to Acoustic,
saving records in the civicrm_queue & civicrm_queue_item tables. To process the queue start
the coworker service by opening a shell on the docker image, moving to the drupal folder
and running coworker run

On the right of the search display is a link to acoustic data - this page
fetchs the acoustic data in real time from Acoustic. On dev you will generally
get an error when you go to this link, because the contact is not in acoustic.
However, if you update the snooze date per above the link will load & you
can see that Acoustic now has the date saved.

This includes updates to 2 packages
https://github.com/eileenmcnaughton/omnimail-silverpop/compare/1.3...master
https://github.com/mrmarkfrench/silverpop-php-connector/compare/v1.5.4...master

Bug: T340040

Change-Id: Ia5f95ae2c169e7614b709b5de49c832575ed490a