Improve handling of temporary setting for forgetmeTest
This improves the way we handle the temporary setting in this class.

The code to set the databaseID for the duration of the test was not overriding the
global civicrm_setting - which meant that on our locals it was not actually
'taking' (& probably fails locally as I think our build generally does
set this setting (see https://wikitech.wikimedia.org/wiki/Fundraising/Data_and_Integrated_Processes/Acoustic_Integration#Sandbox)
so the test would fail locally as the sandbox database id is used, rather than the
one set in the test.

Note the changes to the Helper file are just changes to use preferred methods -
Civi::settings()->get() has it's own caching so we don't need an extra
layer

As long as tests passin CI & locally I don't think any other QA needs to be done
Bug: T340040
Change-Id: I91caec0918acf54537d925d8ff656ab959d9677a

Change-Id: I66efa37f23fe90f673d49c7c36b6fc4443588fc7