Add search display for viewing emails
This adds the search display to view the emails & edit the snooze date. I expect our
first UI for snooze will be this search + a form with filtering at a url and / or
a dashlet. We might look into whether we can get it onto the contact summary.

We don't want to make this functionality obvious to the users yet though as
we need to get coworker running before the syncs will happen and
we probably want to work with one or more donor services agents to test.

In the short term this is probably a step towards where we want to go
and it will help fr-tech to work with the snooze data.

The search is visible at https://wmff.localhost:32353/civicrm/search#/display/snooze/snooze_table
Bug: T340040
Change-Id: Id4159a453e613df952db13fa8b9ecf186e53e5c9

Change-Id: Id7efcbb4696a281b18374e27c67db17ac5524f6c