This patch would go upstream, pushing up here to discuss if this is a potential solution for this unique problem.
Test instructions can be found here: https://lab.civicrm.org/dev/core/-/issues/4467

The bug seemed to be caused by an incorrect use of the array_intersect_key php method. This led to the creation of new fields whenever an update is made to the mapping, causing the data to be inconsistent.
Rearranging the parameters of the method fixes this, here's an example: https://civicrm.frdev.wikimedia.org/civicrm/api4#/explorer/MappingField/get?where=%5B%5B%22mapping_id%22,%22%3D%22,%22212%22%5D%5D. The fields with value '1' or 1 are being set by because instead of the actual values, the method is setting the fields to 'true'. Can be tested with the test instruction above.

Suggestion: The check for the saved field values and the import should be done only on the CRM_Contribute_Import_Form_Preview form, currently it is done in both the CRM_Contribute_Import_Form_MapField and CRM_Contribute_Import_Form_Preview.

Bug: T342722
Change-Id: I1220d00641d6050a8a836d3b354dcdaceec88b0b