Trial shorter Redis cache
This reduces the TTL on our Civi Redis cache from 6 hours to one hour.

At this stage the goal is to just test the impact.

The Redis cache is used for storing metadata and also search results from
UI searches. This latter is referred to as the prevnext cache and is
the more significant usage. Unlike the metadata it will not 'just
reload' as it reflects a specific user's actions through the UI.

The prevnext cache basically stores all the search results from a given
UI search, allowing paging & actions to be taken on it. It seems
fairly safe to assume that a user would NOT be working on the same
search results for longer than an hour. The only scenario Tim thought
of was when someone steps away from their browser for an hour
https://lab.civicrm.org/dev/core/-/issues/4501#note_150102

When I checked recently there were 45k keys in the prevnext cache which
seems like a lot. I can't see this being a full fix for the cache issue
in the gitlab but I would like to deploy it & see

1) the impact it has on the number of keys in the cache
2) the impact it has on the speed of deleting contacts
3) whether users notice - either positively (speed) or
negatively (weird stuff)

Bug: T247347
Change-Id: Iaa6362faebc973aee57758a7c30031b2cbd75aeb