Redis fix - only remove contacts from sub-keys if they are in the main key
This implements bullet point 2 from
https://lab.civicrm.org/dev/core/-/issues/4501#note_150103

The logic is that the contact is only added to 'sel' if they are
in the search and SELECTED (sel = first bit of selected) and only
in 'data' if they are in the search & extra info has been loaded.
In other words we don't need to do those removes if they are
not in 'all'

I checked this out by stepping through the code - ie do a search, find
a contact & delete it. The first time around they will exist in
at least one prevnext key & deleted with be 1. However, if you then
permanently delete them deleted is 0 the second time. Hence the second
2 cache delete hits are not required.

Probably these would normally be cheap - but as we have a current issue with keys
not having expired they have built up to a crazy level (35k) so this would
elinated 20k of cache hits. I want to address the reason they are building
up but deploying this while they are still high will allow me to see the impact
when there ARE a lot of caches to delete from
Bug: T247347
Change-Id: Ib81e8b959dfb2fe02e711652729a47a0b15dcace

Change-Id: I7d0a7a593d68f0cc008f8490f957da10244a4082