Fix MailingJob lookup
I discovered that we are creating new mailings all the time for
thank yous because, on stepping through the test testSendThankYouAddCiviMailActivity
I found that this always returns FALSE

This manefests in our thank_you log with messages like
civicrm.wmf.INFO: thank_you: Could not find mail
ing record for thank_you with name thank_you.en.html and revision 20210323

They happen at the start of the job but not later as
the value is then in a static
I'm not sure this is a complete fix but it is a required fix as
the fetch() method does not return anything when used without having
called executeQuery first. The correct method here is find()

Note the fetch after the find IS working - I just changed it too
for consistency.

This is some old awful code but I'm familiarising myself with it
before thinking about how we might change it...
Bug: T340064

Change-Id: I0513b3e167f1930336ac7c84a1ee313109249c91