Stop creating multiple MailingJob records per mailing
The point of the MailingJob records is to batch sending of Mailings for various mail server
related reasons.

Since we are not sending these emails via CiviMail scheduling there in no
benefit. However, it does cause significant query performance problems - ie
the query on the dashboard (see https://phabricator.wikimedia.org/T335071)
joins the mailing job table to get fields that really belong on the mailing
table (see https://lab.civicrm.org/dev/core/-/issues/4567)

This fixes it so that we stop creating new MailingJob records & instead
keep re-using the same one.

It will need a data cleanup - but we need to think about how long
we need to respect the VERP data for - or alter what it looks for per the
above phab. More reading on this at https://phabricator.wikimedia.org/T335296#9159155

Bug: T335071
Bug: T346194

Change-Id: I30d1eb99541b4627c895e06de11dfea006898754