Minor update using QueueRunner.
This updates a handful of the rows with tax_amount = NULL
to tax_amount = 0.

As with https://gerrit.wikimedia.org/r/c/wikimedia/fundraising/crm/+/961255
the point of this script is more about setting up
the mechanism to have coworker process our updates, taking
into account our busy checks than it is to get the code updates run
through, hence the code only processes 10k rows.

My thinking is we would deploy this & run it before using
the same mechanism to do the meatier updates in

https://phabricator.wikimedia.org/T344012

Note that for testing
1) Run alter table civicrm_contribution modify total_amount decimal(20,2) not null
2) Run UPDATE civicrm_contribution SET tax_amount = NULL;
3) Run drush @wmff cvapi Extension.upgrade
4) Navigate to wmff/drupal and call coworker run

The expecation is that after coworker has had time all rows
in civicrm_contribution will have a tax_amount of 0 and
the item from civicrm_queue_item should be removed again
but the civicrm_queue row will hang around & we will re-use next time

Bug: T344012

Change-Id: Ib42e5755316b7d4c6aada94fe2a074327368aab5