Update community gift & Benefactor gift
These are being phased out. We update all from 1 July to Online Gift.

Note that there is probably an INSERT version of this to do as a follow up still
but trying to break them into separate steps for deployment as we work through
getting used to our new method

Bug: T344012
Change-Id: I0f0ce4cbd2825a2f5db8dedc181c60d49b18bd59