UPI: stop setting early next_sched_contribution_date
We have been advancing the next_sched_contribution_date for UPI
donations because we need to send a pre-notification a day early
to actually get the second installment one month after the initial
donation. Turns out that a mismatched cycle_day and next_sched
date is really bad for donors with cycle_day on the 1st, as it
leads to us sending prenotifications on the last day of the same
month they made the initial donation in, and then sending ANOTHER
prenotification on the first day of the subsequent month, so
we charge them twice in two days.

This patch leaves us with the simplest code, but donation
histories will show all recurring donations on a day-of-month
one day later than the initial donation.

To investigate: do we actually need a different queue consumer
for UPI now, or can we just send messages to the regular recurring
queue?

Alternative to
I34fb32dafb446b4bce8410ef452ae65a3a334d7f and
I66d545f11cbc36684af208e42bebbe1c8889e0fd

Bug: T348258
Change-Id: Ie07ba6ec2811ff737d381a2d3d7c83ac0c8ece7a