Insert step (with special before magic) to remove old queue items
UPDATE - the magic weight was not being set in the new jobs before -
that won't matter in terms of order

Original comment - We are busily (at least on staging) updating queue items but
I think we can just remove some - this inserts a job to do that
with a lower weight than the existing running job - so it should push in.

I updated wikitech
https://wikitech.wikimedia.org/wiki/Fundraising/Internal-facing/CiviCRM#CiviMail_records
with some relevant notes & the linked CiviCRM core docs.

Note this deletes mailing event records back to a year ago.
We could be more ruthless but it will probably take forever to get to
this point anyway so prioritising getting stuff
started & we can discuss whether 3 months or 6 months makes more sense.

We haven't really looked at the bounce records - although the
new activities will surface them to users who might have thoughts

Bug: T335070
Change-Id: I9c6bfee27edc4e75cdd19ffc6f134bcafa6d3a0e