Add Fundraiseup platform fees to fees - Update test.
To confirm values, search for the invoice_id in the fundraiseup sandbox page. For example, the trxn in the tests with invoice_id: DXLKVGQU, can be found here https://dashboard.fundraiseup.com/AVLMPSRU/test/donations/DXLKVGQU. Sum the processing fee and platform fee to see if it equals the value in the fee.

Bug: T346620
Change-Id: I98a093b518c9c31356125ce54f4d6c2127af238e