Update calculation of WMF Donor fields to exclude fields for removal
This will update our dev / CI environments & will allow us to regenerate
triggers sql for wmf_donor which excludes these fields

No change will take place on live until we pro-actively make change
- ie re-roll & update triggers & once loaded we can add sql to
drop the fields. Getting the triggers regenerted early will
allow us to do that early in the maintenance week with the other change
happening later in the week (which we wouldn't want to merge
in advance of iminent deployment)

Note I had to update tests but for testMakeLesserRefund
I dropped the year total check as it was gonna take
a big revisiting of test values now we
can't test for year-1. I figured that was not
the main part of the text

Bug: T347724
Change-Id: I27221484d2cf6edeac3d9d8f5988314d6dc6595c