Remove failed attempt at re-saving weight
I think this weight is failing to save & hence causing the job
to fail to requeue. I think I was working on it at the end of
the day & didn't realise when I restarted that I had failed to
solve it & pushed it up.

Weight is a nice to have. I might keep trying to get it to work but
I realised I got my db in a state jumping between tasks so trying to
just get it 'back on track'. I have re-queued one task, the one
I was trying to get to run first by manipulating weight.
Will move onto the others after
Bug: T335070

Change-Id: Iac7385eae88e0c17450fad2bba75def82e14e862